[contracts] Widen an interior pointer, and skip on the place written
Two frame targets the contract named reached a replaced call site
without being havocked, both silently: the ensures then contradicted the
caller's pre-call memory, the path died and everything after the call
passed vacuously.

- `&b[2]` is not the decay a compiler produces, but it names a place
  inside b that the callee writes forward from, and the array is the
  only object to widen to. Any constant index now widens, which
  over-havocs the elements before it -- that loses the caller
  information rather than granting it any, and it is what a decayed `b`
  already did.
- the value-set skip is about assigning a pointer, not about writing
  through one, so it is tested on the place actually written: after
  widening, and after following a pointer parameter. Tested earlier it
  swallowed the whole frame. `--add-symex-value-sets` is default-enabled
  by `--k-induction`, `--inductive-step` and `--loop-invariant`.

Both were KNOWNBUG in the previous commit and are now CORE.
