+++ regression/function_contract/github_6961_assigns_ptr_var_valuesets_fail/test.desc
@@ -1,4 +1,4 @@
-KNOWNBUG
+CORE
 main.c
 --replace-call-with-contract clr --unwind 10 --add-symex-value-sets
 ^VERIFICATION FAILED$
